Interpolate environment variables.
When the last argument of `local::lib->build_environment_vars_for` is
`0`, which equals to `local::lib->LITERAL_ENV`, the new sub-shell of
`plenv use` will not expand `$PATH`.

E.g., run `plenv use @MODULE`, then display the value of `PATH` in the
new sub-shell:

    $ echo $PATH
    /home/noname/.plenv/libs/5.18.1@MODULE/bin:$PATH

The `$PATH` part of the output is literal, not expanded.

Fix it by passing `local::lib->INTERPOLATE_ENV` (whose value is `1`).